@@ -5,7 +5,7 @@
     <groupId>org.icatproject</groupId>
     <artifactId>ids.server</artifactId>
     <packaging>war</packaging>
-    <version>1.12.1-SNAPSHOT</version>
+    <version>2.0.0-SNAPSHOT</version>
     <name>IDS Server</name>
 
     <properties>
@@ -73,7 +73,7 @@
         <dependency>
             <groupId>org.icatproject</groupId>
             <artifactId>icat.client</artifactId>
-            <version>4.10.0</version>
+            <version>6.0.0</version>
         </dependency>
 
         <dependency>
@@ -103,9 +103,10 @@
         </dependency>
 
         <dependency>
-            <groupId>javax</groupId>
-            <artifactId>javaee-api</artifactId>
-            <version>7.0</version>
+            <groupId>jakarta.platform</groupId>
+            <artifactId>jakarta.jakartaee-api</artifactId>
+            <version>10.0.0</version>
+            <scope>provided</scope>
         </dependency>
 
         <dependency>
@@ -115,15 +116,16 @@
         </dependency>
 
         <dependency>
-            <groupId>commons-fileupload</groupId>
-            <artifactId>commons-fileupload</artifactId>
-            <version>1.3.3</version>
+            <groupId>com.sun.xml.ws</groupId>
+            <artifactId>jaxws-rt</artifactId>
+            <version>4.0.0</version>
+            <scope>test</scope>
         </dependency>
 
         <dependency>
-            <groupId>org.glassfish</groupId>
-            <artifactId>javax.json</artifactId>
-            <version>1.0.4</version>
+            <groupId>org.eclipse.parsson</groupId>
+            <artifactId>parsson</artifactId>
+            <version>1.1.0</version>
             <scope>test</scope>
         </dependency>
 
@@ -158,7 +160,7 @@
             <plugin>
                 <groupId>com.qmino</groupId>
                 <artifactId>miredot-plugin</artifactId>
-                <version>1.6.2</version>
+                <version>2.4.1-Java11</version>
                 <executions>
                     <execution>
                         <id>miredot</id>
@@ -169,7 +171,7 @@
                 </executions>
                 <configuration>
                     <licence>
-                        cHJvamVjdHxvcmcuaWNhdHByb2plY3QuaWRzLnNlcnZlcnwyMDIyLTA0LTAxfGZhbHNlfC0xI01Dd0NGQjFzVFhsRS9CL1llTklZWndCZnB2UHlramcrQWhSYW5NR2JOS2craFZMUzJwRzZtcnZ3WXlEMzB3PT0=
+                        cHJvamVjdHxvcmcuaWNhdHByb2plY3QuaWRzLnNlcnZlcnwyMDI2LTA1LTE4fGZhbHNlfC0xI01Dd0NGQnVZbjdlYjEzWXluWFpLa1VRNnA2VDZBVXRnQWhSUlZpMUJTald2KzFobUV3QzdmQko2dE44cld3PT0=
                     </licence>
                     <output>
                         <html>
@@ -263,10 +265,9 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.3</version>
+                <version>3.10.1</version>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
+                    <release>11</release>
                 </configuration>
             </plugin>
 
@@ -318,7 +319,7 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-war-plugin</artifactId>
-                <version>3.0.0</version>
+                <version>3.4.0</version>
             </plugin>
 
             <plugin>
